a new atom should not contain a pointer to the parent
From my experience with doublely linked trees/linked lists, a new node should not contain a pointer to a parent. This pointer should be added explicitly from the parent's "add child" method. (e.g. https://doc.rust-lang.org/src/alloc/collections/linked_list.rs.html#120 )

In this case, for example, someone might assume that an atom can be added to a chain by providing a Some(pointer) as the first argument. However, this fails, because in the `new` method the new atom is not pushed to the parent's Vec of atoms.